requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
click==8.1.8
colorama==0.4.6
Django==3.2.25
django-crispy-forms==2.0
djangorestframework==3.15.1
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.5
MarkupSafe==2.1.5
numpy==1.21.6
pandas==1.1.5
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
